+++ src/app/search/search-filters.ts
@@ -123,7 +123,9 @@ export function buildSearchConfig(
       powermod: ['powermod', 'haspowermod'],
       complete: ['goldborder', 'yellowborder', 'complete'],
       masterwork: ['masterwork', 'masterworks'],
-      hasShader: ['shaded', 'hasshader']
+      hasShader: ['shaded', 'hasshader'],
+      prophecy: ['prophecy'],
+      ikelos: ['ikelos']
     });
   }
 
@@ -248,6 +250,28 @@ export function searchFilters(
     "Ship",
     "Ships",
     "ClanBanners"
+  ]);
+
+  const prophecyHash = new Set([
+    472169727,
+    3991544423,
+    3285365666,
+    161537636,
+    2091737595,
+    3991544422,
+    3285365667,
+    161537637,
+    3188460622,
+    1490571337,
+    2248667690, // perfect paradox
+    573576346   // sagira shell
+  ]);
+
+  const ikelosHash = new Set([
+    847450546,
+    1723472487,
+    1887808042,
+    3866356643
   ]);
 
   // This refactored method filters items by stats
@@ -867,6 +891,12 @@ export function searchFilters(
       },
       armor(item: DimItem) {
         return item.bucket && item.bucket.sort === 'Armor';
+      },
+      prophecy(item: D2Item) {
+        return prophecyHash.has(item.hash);
+      },
+      ikelos(item: D2Item) {
+        return ikelosHash.has(item.hash);
       },
       cosmetic(item: DimItem) {
         return cosmeticTypes.has(item.type);